api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.5.0
websockets==13.1
httpx==0.27.2
firebase-admin==6.6.0
google-cloud-storage==2.18.2
sarvamai>=0.1.0
scrapling>=0.4.0
trafilatura>=1.6.0
